refactor: 136 - move products-search into different folder
